pyproject.toml
[project]
name = "harness-lab"
version = "0.1.0"
description = "Educational coding-agent harness using PydanticAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "logfire>=4.32.0",
    "pydantic-ai>=1.80.0",
    "pydantic-ai-slim[duckduckgo,web-fetch]>=1.80.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "rich>=15.0.0",
    "sqlmodel>=0.0.38",
    "sentence-transformers>=5.4.0",
    "sqlite-vec>=0.1.9",
    "httpx>=0.28.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
harness = "src.client.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pydantic-evals>=1.80.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
    "rumdl>=0.1.70",
    "taskipy>=1.14.1",
    "ty>=0.0.29",
]
